#7 Switching back to RGB-48 this limits panel size to 64x64 for GEN 3. TLC5958 is in theory capable of 64x128 see #7 for details. (LEDs only have 12 bits of color depth and multiplex lowers this.)

+++ LED_Matrix/include/Matrix/config.h
@@ -23,7 +23,7 @@
     #include <math.h>
     constexpr int PWM_bits = round(log2((double) MAX_RGB_LED_STEPS / MULTIPLEX));
     
-    typedef uint8_t test[2 * MULTIPLEX][COLUMNS][3];
+    typedef uint16_t test[2 * MULTIPLEX][COLUMNS][3];
     
     typedef union {
         test data;
